Consolidate time code components.
