--- src/GoatSlips/Client/src/components/ManageTimeCodes/ManageProjectCodes.module.scss
@@ -1,35 +1,0 @@
-$newProjectHeight: 74px;
-$newProjectMargin: 5px;
-
-.newProjectContainer {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: $newProjectHeight;
-    margin-bottom: $newProjectMargin;
-    border: 1px solid grey;
-    border-radius: 5px;
-}
-
-.projectName {
-    width: 250px !important;
-}
-
-.addNewProject {
-    margin-left: 0.7em !important;
-}
-
-.tableContainer {
-    $marginBottom: 20px;
-    height: calc(100% - $newProjectHeight - $newProjectMargin - $marginBottom);
-    margin-bottom: $marginBottom;
-    min-width: 1000px;
-}
-
-.taskCell {
-    width: 300px;
-}
-
-.buttonCell {
-    width: 150px;
-}
