--- src/GoatSlips/Client/src/components/ManageTimeCodes/ManageTaskCodes.module.scss
@@ -1,31 +1,0 @@
-$newTaskHeight: 74px;
-$newTaskMargin: 5px;
-
-.newTaskContainer {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: $newTaskHeight;
-    margin-bottom: $newTaskMargin;
-    border: 1px solid grey;
-    border-radius: 5px;
-}
-
-.TaskName {
-    width: 250px !important;
-}
-
-.addNewTask {
-    margin-left: 0.7em !important;
-}
-
-.tableContainer {
-    $marginBottom: 20px;
-    height: calc(100% - $newTaskHeight - $newTaskMargin - $marginBottom);
-    margin-bottom: $marginBottom;
-    min-width: 500px;
-}
-
-.buttonCell {
-    width: 0;
-}
