+++ src/GoatSlips/Client/src/components/pages/ManageLaborCodes.module.scss
@@ -1,5 +1,25 @@
+$headerHeight: 90px;
 $newLaborCodeHeight: 74px;
 $newLaborCodeMargin: 5px;
+
+.pageContainer {
+    height: 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.header {
+    height: $headerHeight;
+    font-size: 30px;
+    display: flex;
+    align-items: center;
+}
+
+.codesList {
+    max-height: calc(100% - ($headerHeight * 2));
+}
 
 .newLaborCodeContainer {
     display: flex;
