+++ src/GoatSlips/Client/src/components/pages/ManageProjects.tsx
@@ -1,12 +1,24 @@
-import { Alert } from '@mui/material';
-import React, { useState } from 'react';
+import {
+    Alert,
+    Button,
+    Paper,
+    Table,
+    TableBody,
+    TableCell,
+    TableContainer,
+    TableHead,
+    TableRow,
+    TextField,
+} from '@mui/material';
+import { useState } from 'react';
+import { fetchPostResponse } from '../../helpers/fetchFunctions';
 import { AlertMessage } from '../../types/AlertMessage';
 import { DropdownOption } from '../../types/DropdownOption';
 import { Project } from '../../types/Project';
-import { ManageProjectCodes } from '../ManageTimeCodes/ManageProjectCodes';
+import { ExistingProjectRow } from '../ManageTimeCodes/ExistingProjectRow';
 import classes from './ManageProjects.module.scss';
 
-interface ManageTimeCodesProps {
+interface ManageProjectsProps {
     projects: Project[];
     tasks: DropdownOption[];
     taskMap: Map<number, string>;
@@ -18,7 +30,7 @@ interface ManageTimeCodesProps {
     managedProjectIds: Set<number>;
 }
 
-export const ManageProjects: React.FC<ManageTimeCodesProps> = (props: ManageTimeCodesProps) => {
+export const ManageProjects: React.FC<ManageProjectsProps> = (props: ManageProjectsProps) => {
     const {
         projects,
         tasks,
@@ -31,7 +43,83 @@ export const ManageProjects: React.FC<ManageTimeCodesProps> = (props: ManageTime
         managedProjectIds,
     } = props;
 
+    const [newProjectName, setNewProjectName] = useState('');
     const [alertMessage, setAlertMessage] = useState<AlertMessage | null>(null);
+
+    const handleNewProjectNameChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
+        const value = event.target.value;
+
+        setNewProjectName(value);
+    };
+
+    const createProject = async () => {
+        await fetchPostResponse('Project/Create', { projectName: newProjectName });
+        setNewProjectName('');
+        await fetchProjects();
+        await fetchTasksAllowed();
+    };
+
+    const fetchProjectsAndFavorites = async () => {
+        await fetchProjects();
+        await fetchFavorites();
+    };
+
+    const existingProjectElements = projects.map((project: Project) => {
+        if (!isAdmin && !managedProjectIds.has(project.id)) {
+            return null;
+        }
+        return (
+            <ExistingProjectRow
+                allTasks={tasks}
+                tasksAllowed={tasksAllowedForProjects.get(project.id) ?? []}
+                taskMap={taskMap}
+                key={`project${project.id}`}
+                project={project}
+                fetchProjects={fetchProjectsAndFavorites}
+                fetchTasksAllowed={fetchTasksAllowed}
+                setError={(message: string) => setAlertMessage({ message, severity: 'error' })}
+                setSuccess={(message: string) => setAlertMessage({ message, severity: 'success' })}
+            />
+        );
+    });
+
+    const getProjectsList = () => {
+        return (
+            <>
+                <div className={classes.newProjectContainer}>
+                    <TextField
+                        className={classes.projectName}
+                        label="New Project"
+                        variant="outlined"
+                        value={newProjectName}
+                        onChange={handleNewProjectNameChange}
+                    />
+                    <Button
+                        className={classes.addNewProject}
+                        variant="contained"
+                        color="success"
+                        disabled={newProjectName === ''}
+                        onClick={createProject}
+                    >
+                        Add
+                    </Button>
+                </div>
+                <TableContainer component={Paper} className={classes.tableContainer}>
+                    <Table stickyHeader>
+                        <TableHead>
+                            <TableRow>
+                                <TableCell>Name</TableCell>
+                                <TableCell className={classes.taskCell}>Allowed Tasks</TableCell>
+                                <TableCell>Rate</TableCell>
+                                <TableCell className={classes.buttonCell} />
+                            </TableRow>
+                        </TableHead>
+                        <TableBody>{existingProjectElements}</TableBody>
+                    </Table>
+                </TableContainer>
+            </>
+        );
+    };
 
     const getAlert = () => {
         if (alertMessage === null) {
@@ -45,20 +133,7 @@ export const ManageProjects: React.FC<ManageTimeCodesProps> = (props: ManageTime
         <div className={classes.pageContainer}>
             <div className={classes.header}>Manage Projects</div>
             {getAlert()}
-            <div className={classes.codesList}>
-                <ManageProjectCodes
-                    projects={projects}
-                    tasks={tasks}
-                    taskMap={taskMap}
-                    tasksAllowedForProjects={tasksAllowedForProjects}
-                    fetchProjects={fetchProjects}
-                    fetchTasksAllowed={fetchTasksAllowed}
-                    fetchFavorites={fetchFavorites}
-                    setAlertMessage={setAlertMessage}
-                    isAdmin={isAdmin}
-                    managedProjectIds={managedProjectIds}
-                />
-            </div>
+            <div className={classes.codesList}>{getProjectsList()}</div>
         </div>
     );
 };
